pyproject.toml
[project]
name = "ai_agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Bhavika",email = "bhavika.rawat.24@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
